requirements.txt
certifi==2020.12.5
chardet==4.0.0
color-transfer==0.1
idna==2.10
imutils==0.5.3
numpy==1.19.2
opencv-contrib-python==4.4.0.44
python-consul==1.1.0
requests==2.25.1
six==1.15.0
urllib3==1.26.3
